Organize assets and stylesheets
