+++ src/assets/images/equilateral.jpeg
