+++ src/assets/stylesheets/_variables.scss
