+++ src/assets/stylesheets/application.scss
@@ -1,0 +1,16 @@
+@import './partials/BackButton.scss';
+@import './partials/DevSummary.scss';
+@import './partials/Discography.scss';
+@import './partials/DiscographyItem.scss';
+@import './partials/Email.scss';
+@import './partials/Github.scss';
+@import './partials/Header.scss';
+@import './partials/Headshot.scss';
+@import './partials/HomeButton.scss';
+@import './partials/Instagram.scss';
+@import './partials/Licensing.scss';
+@import './partials/MusicianSummary.scss';
+@import './partials/OutArrow.scss';
+@import './partials/Portfolio.scss';
+@import './partials/Press.scss';
+@import './partials/ProjectSection.scss';
