+++ src/assets/stylesheets/partials/Email.scss
