+++ src/assets/stylesheets/partials/Headshot.scss
@@ -1,4 +1,4 @@
-@import '../style/variables';
+@import '../variables';
 
 .paintedHeadshot {
     border-radius: 25px;
@@ -10,8 +10,6 @@
     width: 450px;
     height: 450px;
     border: 10px solid $transparent-background;
-
-
 
     &:hover{
       animation: pulse 5s infinite;
