+++ src/assets/stylesheets/partials/OutArrow.scss
