+++ src/assets/stylesheets/partials/ProjectSection.scss
