+++ src/index.scss
@@ -1,22 +1,23 @@
-@import './style/App';
-@import './style/variables';
-@import './style/reset';
-@import './style/extends';
-@import './Headshot/Headshot.scss';
-@import './Discography/Discography.scss';
-@import './Licensing/Licensing.scss';
-@import './Portfolio/Portfolio.scss';
-@import './DevSummary/DevSummary.scss';
-@import './Press/Press.scss';
-@import './Twitter/Twitter.scss';
-@import './Instagram/Instagram.scss';
-@import './Github/Github.scss';
-@import './Email/Email.scss';
-@import './Header/Header.scss';
-@import './Arrow/Arrow.scss';
-@import './MusicianSummary/MusicianSummary.scss';
-@import './BackButton/BackButton.scss';
-@import './OutArrow/OutArrow.scss';
+@import './assets/stylesheets/App';
+@import './assets/stylesheets/variables';
+@import './assets/stylesheets/reset';
+@import './assets/stylesheets/extends';
+// @import './Headshot/Headshot.scss';
+// @import './Discography/Discography.scss';
+// @import './Licensing/Licensing.scss';
+// @import './Portfolio/Portfolio.scss';
+// @import './DevSummary/DevSummary.scss';
+// @import './Press/Press.scss';
+// @import './Twitter/Twitter.scss';
+// @import './Instagram/Instagram.scss';
+// @import './Github/Github.scss';
+// @import './Email/Email.scss';
+// @import './Header/Header.scss';
+// @import './Arrow/Arrow.scss';
+// @import './MusicianSummary/MusicianSummary.scss';
+// @import './BackButton/BackButton.scss';
+// @import './OutArrow/OutArrow.scss';
+@import './assets/stylesheets/application.scss';
 
 body, html {
     background-color: $primary-background-color;
